+++ ai-services/assets/worker/podman/templates/worker.yaml.tmpl
@@ -10,7 +10,6 @@ metadata:
     ai-services.io/volume: "podman-auth-secret,worker-tls"
   annotations:
     run.oci.keep_original_groups: "1"
-    
 spec:
   restartPolicy: always
   containers:
