+++ ai-services/internal/pkg/worker/gateway/pki.go
@@ -72,7 +72,7 @@ func generateCA() (*ecdsa.PrivateKey, *x509.Certificate, []byte, error) {
 	caSerial, _ := rand.Int(rand.Reader, new(big.Int).Lsh(big.NewInt(1), serialBitSize))
 	caTemplate := &x509.Certificate{
 		SerialNumber:          caSerial,
-		Subject:               pkix.Name{CommonName: "ai-services-worker-ca"},
+		Subject:               pkix.Name{CommonName: "catalog-worker-ca"},
 		NotBefore:             time.Now(),
 		NotAfter:              time.Now().Add(caTTL),
 		IsCA:                  true,
@@ -100,7 +100,7 @@ func generateServerCert(caCert *x509.Certificate, caKey *ecdsa.PrivateKey) (*ecd
 	srvSerial, _ := rand.Int(rand.Reader, new(big.Int).Lsh(big.NewInt(1), serialBitSize))
 	srvTemplate := &x509.Certificate{
 		SerialNumber: srvSerial,
-		Subject:      pkix.Name{CommonName: "WorkerGateway"},
+		Subject:      pkix.Name{CommonName: "Catalog"},
 		DNSNames:     []string{workerconstants.GatewayServerName},
 		NotBefore:    time.Now(),
 		NotAfter:     time.Now().Add(serverCertTTL),
@@ -224,42 +224,3 @@ func fileExists(path string) bool {
 	return err == nil
 }
 
-// signWorkerCSR validates and signs a PEM-encoded CSR using the gateway CA,
-// returning the signed certificate and CA certificate as PEM bytes together with
-// the cert's expiry time.  All crypto details stay in this file, keeping
-// gateway.Register focused on the RPC flow.
-//
-// workerName is embedded as the cert CN so the worker can recover its registered
-// name from the cert on reconnect without any additional state file.
-func signWorkerCSR(csrPEM []byte, workerName string, caCert *x509.Certificate, caKey *ecdsa.PrivateKey) (certPEM, caCertPEM []byte, notAfter time.Time, err error) {
-	block, _ := pem.Decode(csrPEM)
-	if block == nil || block.Type != "CERTIFICATE REQUEST" {
-		return nil, nil, time.Time{}, fmt.Errorf("malformed CSR: not a valid PEM CERTIFICATE REQUEST block")
-	}
-	csr, err := x509.ParseCertificateRequest(block.Bytes)
-	if err != nil {
-		return nil, nil, time.Time{}, fmt.Errorf("parse CSR: %w", err)
-	}
-	if err := csr.CheckSignature(); err != nil {
-		return nil, nil, time.Time{}, fmt.Errorf("invalid CSR signature: %w", err)
-	}
-
-	serial, _ := rand.Int(rand.Reader, new(big.Int).Lsh(big.NewInt(1), serialBitSize))
-	notAfter = time.Now().Add(workerCertTTL)
-	tmpl := &x509.Certificate{
-		SerialNumber: serial,
-		Subject:      pkix.Name{CommonName: workerName, Organization: csr.Subject.Organization},
-		NotBefore:    time.Now(),
-		NotAfter:     notAfter,
-		ExtKeyUsage:  []x509.ExtKeyUsage{x509.ExtKeyUsageClientAuth},
-		KeyUsage:     x509.KeyUsageDigitalSignature,
-	}
-	signedDER, err := x509.CreateCertificate(rand.Reader, tmpl, caCert, csr.PublicKey, caKey)
-	if err != nil {
-		return nil, nil, time.Time{}, fmt.Errorf("sign certificate: %w", err)
-	}
-	certPEM = pem.EncodeToMemory(&pem.Block{Type: "CERTIFICATE", Bytes: signedDER})
-	caCertPEM = pem.EncodeToMemory(&pem.Block{Type: "CERTIFICATE", Bytes: caCert.Raw})
-
-	return certPEM, caCertPEM, notAfter, nil
-}
